refactor: move useCurrentPage to modules
